Do not hash stack past length
As an optimization, we don't zero the scratch stack every single time
and simply set its length to zero.

This commit ensures that we don't try to read these garbage values while
hashing because then the hash for two stacktraces that are equivalent
will be different reducing the space in the hashmap where they are
inserted.