Fix executable mapping finder
Rust's binary search yields a result: Ok() if the exact element was
found, or an Err() with the insertion index if no exact matches were
found.

This was obviously buggy before. Additionally there were no checks for
the mappings being sorted, and while they typically are we have no
guarantees for this. As the vector is small enough (250 elements atm)
let's do a linear search and in the future revisit this if needed.

Test Plan
=========

Collected a bunch of profiles and they all looked good.