Unwind kernel stacks and add initial kernel symbolizer
Right now we store all the kernel symbols in memory, which can be quite
heavy weight (50-150 MB in many machines). This will be reworked to use
a DB such as sqlite

Test Plan
=========

Checked a bunch of profiles, the kernel stacks are in the right spots
and contain frames that are sensible both in order and in "weight".