ci: Run tests
We were not running tests in ci, so fix this, also add a step just to
enter the nix dev environment so the we know how long does it take to
pull / build our native dependencies and tooling.

Also changing `check` -> `build` as the CLI integration tests require
the debug binary to be built. This is a bit janky and we should have
better integration with the build system. We'll change this later on.

Test Plan
=========

ci :)