build: Static build with glibc
This is a first iteration for statically built lightswitch. Using glibc is
not ideal for most statically linked applications, as it might dynamically
load other code that dynamically links libc, resulting in two copies of
libc linked in the same address space.

Other possible approaches are:

- Using musl as the libc: tried this approach but couldn't get it to
  link after couple of hours of trying.
- Build an appimage: we could do this with dynamic or static libraries,
  but the resulting executable presents itself as a self-contained statically built binary.

Notes: I tried to make statically linked zstd work, but after many many
hours of trying to figure out why it would not find the symbols, I
decided to disable it. Seems like not that many compilers aren't compressing sections
with zstd yet, but ideally in the future we migth want to enable this.
This is also disabled in the vendored elfutils that libbpf-sys exposes
[0].

After statically compiling the project, it gets uploaded as an artifact.
This should help us debug issues in the future and bisect problems
faster.

- [0]: https://github.com/libbpf/libbpf-sys/commit/2640a5f8daaf8cb6b7cd02ea3acf7daf37ef5a89#diff-d0d98998092552a1d3259338c2c71e118a5b8343dd4703c0c7f552ada7f9cb42R279

Test Plan
=========