Use libbpf's CPU count
The `num_cpus` crate might report a different number of CPUs that libbpf
does. This is problematic in various ways, such as when updating per-CPU
BPF maps that require one entry per CPU. Internally, libbpf-rs relies on
libbpf's CPU count call to ensure that the right number of items is passed.