Object file improvements
Do not memory map and unmap the same object file 3 times, and make the
code a bit cleaner.

In order to make this code work I had to create a static lifetime, and
then manually drop it, this is not ideal, but it's the best I could do
after many hours.

Also, handle errors a bit more gracefully, for example, before this
commit, 32 bit binaries can cause panics.

Test Plan
==========

Ran for a while and also under valgrind.