Rework known process detection
We don't need a separate map anymore. As BPF tracing programs require
hash maps to be preallocated, this will save some memory.

Something we might want to do in the future is tune the maximum entries
in the mappings BPF map.

Additionally, this also cleans up some of the implementations for the
BPF bindings. I love bindgen!!

Test Plan
=========

Ran locally without any issues or errors.