Rate limit unwinder events
This is needed because otherwise the perf buffer can drop events if it
gets too full too quickly. This can happen for a variety of reasons, one
of them that a hot process doesn't have any unwind info so it keeps on
sending the events to userspace.

There's no advantage of sending the event multiple times, we should
always monitor this perf buffer to ensure we don't drop events though.

Test Plan
=========

Ran locally for a while without issues, can see rate limits in the logs.