unwinder/native: Log vDSO segments
Unwinding through vDSO regions is not supported yet. We need to open the
non-file memory mapping and retrieve its unwind info, but until it's
implemented let's log with the right message.

The mapping types will probably disappear, that's why I am not bothering
with extracting them into a constant or enum.