Improve logging (#40)
Fix the log verbosity for the userspace program to let the user select
the level they want to filter by. This also replaces the previous description
which wasn't clear enough.

Add two other flags to set logging of the libbpf library which also
contains verifier errors, as well as for the BPF programs themselves. By
default they are turned off to reduce noise and performance overhead.

Test Plan
=========

Ensured the logging was off by default and could be enabled passing the
right flags.