unwinder: Fix assumption on thread's initial registers
The x86_64 ABI mandates that the bottom frame of processes must have
their frame pointer (rbp) set to zero. We check for this when we detect
and end of stacktrace condition, which is correcto for processes but
it's not for threads as their initial frame pointer value is not
specified.

Test Plan
=========

Ran for a while without issues. Unwinding Node.js applications cause the
log to show up, this needs more investigation, but in general we should
always use frame pointers to unwind Node. We will revisit this once we
add proper support for this runtime.