Improve main executable detection
The logic here is a bit flimsy and should be reworked, but let's at
least mark first executables as the "main" ones.

More context in the comment but overall this is not always correct, in
particular Rust + static musl is broken due how the loader operates for
that case.

We should understand the kernel and loader's logic in detail and fully
replicate it here.

Test Plan
=========

Ran locally for several minutes without issues. The profilers looked
good (see PR) and 0 unexpected errors were emitted.