Implement unwind information reset
This implmementation has several shortcomings, and doesn't have any
rate-limiting in place. We will continue to improve this and have plans
to revisit how we store chunks (as this can be optimized) as well as how
the unwind information is laid out to try to prevent expensive state
resets.

Test Plan
=========

Ran on busy hosts that used all shards. The state was resest and
profiling continued as usual.