Add reference count to object files and fix various bugs
Add reference count to object files. This is not used yet but in the
future this can help us free some internal structures that are no longer
needed.

Additionally, this commit also addresses:

- a bug in how the main executable was determined, which was using the
  object files rather than the mappings, which can cause failures
unwinding in the main executable code regions;
- a bug that would show as processes that had all their executables
  cached already would not be profiled due to a logic bug introduced as
part of some feature that was reworked and is not longer needed;