Introduce custom executable id (#46)
Before this commit the only unique identifier for an executable file or
object file was the build id. As the GNU's build ID might not be present,
we fall back to the sha256 hash of the `.text` section. This works fine but
takes quite a bit of space as we refer to it in every mapping.

By using a 64 bit id derived from the first 8 bytes of the sha256 hash
of the code, we can save some memory, make comparisons and hash the IDs faster,
and helps simplify the code. For example, there's no need for a separate ID in
the native unwind state. There might be a bit more CPU and IO needed to hash
the code, but this might be something we can cache in the future.

This is also useful for future profile formats that require unique 64 bit
IDs for object files.

Also fixes other minor things, such as using the lower hexadecimal representation.

Test Plan
=========

Lots of manual tests, everything seems good.