Aggregate samples in collector and move symbolizer
Before this commit, we were generating profiles, such as flamegraphs
based on a collection of profiles. While there's nothing inherently
wrong about this representation, we were never aggregating the stacks,
which would result in duplicated stack (a sample without a count)
entries. This would be fed directly to the profile renderers and could
potentially cause behaviours that aren't well defined. For example, only
one of the samples (sometimes the first one, in other implementations
the last one) would be used.

Additionally, the collector has no businessin symbolization so this is
now done elsewhere.

Test Plan
=========

Collected a couple of profiles, they all look good.