Add virtual address and file offset for frames
We need both to be able to find memory mappings (*) and the offset in
the object file (what we typically refer to normalized address) to be
able to symbolize an address.

(*) Having to compute mappings in many parts of the code is not ideal
but we'll keep around both normalized and virtual addresses until the
profiler design is more settled.

Test Plan
=========
Manual tests.