build: Optimize all dependenciess in debug mode
Some of the code is excruciatingly slow in debug mode but there are
several more assertions turned on in this mode. Compile dependencies
with the highest level of optimizations to reduce the slow down.

Compile times might be slightly higher but we should not be recompiling
our dependencies very often.