Add integration tests
Starting with C++ binaries. This work leverages the nix build system to
build the test programs [0] and executes them while running the profiler
to ensure the traces are what we expect.

There are some future improvements such as testing other languages and
programs, as well as asserting metadata, and improving the assertion
helper to ensure that the frames are correct by changing it to be a
macro and not converting it to a string

Test Plan
=========
Added tests, CI

[0]: https://github.com/javierhonduco/lightswitch-testprogs