Use `crossbeam_channel` and stop profiling on control+c (#55)
The `crossbeam_channel` crate offers a couple of nice features that the
standard library producer/comsumer channel doesn't have, including a
`select!` macro that lets us poll from the first available channel when
we are polling from several channels. This commit reworks this logic to
unify everything using this macro, resulting in potential fewer waits in
the worst case scenario and cleaner logic.

This work started to add support for SIGINT aka control+c handling,
which is included in this commit. Once the signal is received, profiling
will stop and the profile will be dealt with as soon as the channels are
polled again.

Test Plan
=========

Interrupted profiling with control+C without any issues, made sure it
works for short and long durations to verify there are no regressions.