Fix race while copying `ObjectFileInfo`
There was a fundamental race that could (and did!) happen while copying
/ cloning `ObjectFileInfo`, as we tried to open the file using the
original file path, if the file had been removed, re-opening would fail.
This commit leverages procfs to reopen a file that might already have
been deleted, as long as there are some file descriptors pointing to it,
as the kernel won't effectively remove the file until its reference
count reaches zero.

Test Plan
=========

Added test to exercise this code path. Also run the profiler while
compiling the Linux kernel as this as this triggered the crash in the
past.