Improve handling of deleted executable object files
As before this commit, if lightswitch finds an object file that's been
deleted, which are named as `$original_name (deleted)` we would return
an error and never add the unwind information for any of the object
files. We could see this for example in lightswitch as I assume some
parts of the runtime could load the debuginfo file at runtime and then
remove it at some point.

This commit also slighly improves the code/

Test Plan
=========

Ran while running another `lightswitch` instance, the error is not
showing up anymore.