Various pprof improvements
Send the right profile duration. The profile start time is still the end
time rather than the start time, but this is offset by the profiling
session duration (5s) so it's acceptable as of now.

There are planned changes on the way samples are collected from the
unwinders so the start time will be tackled then.

Test Plan
=========

CI