Initial implementation of debug information manager
Currently, executables are opened as soon as possible to have a hold of
them to be able to symbolize later on. This approach is not ideal
because the number of opened file descriptors can balloon, and this is
done without first checking if the executables contain debug information
at all.

In the future all the local and remote use-cases will transition to
using the debug information manager. This first iteration is quite
simple and lacks a lot of features needed to make this really work. In
this commit it's still disabled by default, and just adds some basic
features to upload debug information to a server, if needed.

Left various TODOs on the planned changes.

Test Plan
=========

CI + manual checks.