perf: Read debug info in streaming / chunks (#138)
Rather than allocating space to store it all in memory. This was done
like this before as during prototyping I did not know if reqwest would
also do this or stream the file.

I've since memory profiled reqwest and checked its source code and it
doesn't keep the whole body in memory.

Test Plan
=========

Manually checked that it continues to work and that the file is read in
chunks with a memory profiler (heaptrack).