Rework executable id generation
The executable is now based on the first 8 bytes of the build id, which
reduces the io and cpu needed to hash the code section from every
binary. This is now only done for executables without GNU or Go build
id. Besides the performance improvement, this will standardise how build
ids are handle once kernel code sections (main executable but also
modules) are processed.

Note: this is a backwards-incompatible change as the executable id will
change for GNU and Go binaries.

Test Plan
=========

Unit tests + ran the profiler for a while w/o issues