Increase debuginfo upload timeout
As it was too slow for larger files. This issue is now more apparent as
the file is read in chunks, which is good to reduce peak memory usage,
but might take slightly longer.

Also now the two http clients are created once, in the constructor, and
the timeout of the two calls are configurable separately.

Test Plan
=========

ci