ci: Cache Rust builds (#160)
This project uses nix to set the developer environment and to build the
container images but for linting or testing the nix provided `cargo` is
used directly. While we could move everything to nix and leverage its
cache I want to make sure that building the project works well in other
environments, which this approach can help with.

This means that on every single CI run all the dependencies have to be
built from scratch, causing elevated build times (~2m in nix vs ~8m
without).

Not 100% sure if this action will work with this project's setup but
we'll see. (update: tested that the cache worked fine)

Test Plan
=========

CI