Find instruction before return address
As we don't want the return address as it points to the next instruction
after the function call. See comments for implementation details for
different architectures.

The kernel unwinder seems to have the same issue, so this is something
that should be adjusted in userspace to avoid using more time in the
unwinder. This will be fixed later on.

Test Plan
=========

Ran a bunch of tests locally and some of the previous off-by-ones while
symbolising are gone.