Fix off-by-one in BPF pages deletion (#179)
The last page was not deleted which resulted in the BPF map to
eventually be full and no more mappings could be added to it. This
commit fixes it by making sure the last page is also deleted, and adding
integration tests to verify that all the necessary BPF maps are cleaned
up.

Test Plan
=========

Added tests and ran for a while whilst spawning many new processes with
different unwind information without issues.