Improve error handling dealing with unwind info
This was discovered thanks to
https://github.com/javierhonduco/lightswitch/commit/37a0dc8de23214b9573a1ff5fd3c1bea13043672.
Basically the current poor error handling was causing data in BPF maps
that should be deleted to be left behind. This surfaced on arm64 as vDSO
sections don't have unwind information and this error code path ran.

Test Plan
=========

The test now passes on arm64.