Use `mmap()` to write unwind info
This commit changes how unwind information is written to the BPF maps
where it's stored. Instead of using batched updates it uses the
`BPF_F_MMAPABLE` option introduced in kernel version 5.5 [0], which was
released ~5y ago, so it's supported by all the kernels we care about.

Saying that the improvement is "good" would be a significant
understatement. It's now 70x faster! Batched BPF updates seem
to be way slower than I expected. Not sure why that's the case.

I couldn't believe the CPU profiles obtained with lightswitch
itself, so it wrote some benchmarks and also ran `strace` and `perf
trace` to make sure that the numbers made sense.

Note that thanks to using `mmap()` reduces the amount of data buffering.
Before this commit, we had to read from disk -> chunks -> bpf maps. Now
the intermediate buffer is no longer needed. So not everything's
directly attributable to using mmap, but it does enable all the
optimizations.

Potentially using larger mmap writes would be faster due to `memcpy`
having optimizations for large memory copies, but this is so fast right
not compared to the previous implementation that I don't think this
would be necessary.

The only downside I can think of is that now the allocations that back
the BPF arrays need to be page aligned, and in case of larger pages,
beyond 4KB there could be some wastage. For the standard page size, this
is not really an issue as unwind information can get quite large
(dozens of MBs is not unheard of).

Test Plan
=========

See PR for benchmarks, profiles, etc.

- [0]: https://github.com/torvalds/linux/commit/fc9702273e2edb90400a34b3be76f7b08fa3344b