unwind_info/pages: Add support for large gaps
Before this commit, large gaps (>=2**16, which is the page size) would
not get their own page. This isn't typically an issue with the unwind
information that most compilers generate in the .eh_frame section, as it
covers every single individual instruction, not leaving any gaps.

When vDSO support was added, this became a potential problem but in
practice the code fit in a single page, so it was a non-issue, until
now for Go support!

Go is compiled with frame pointers so we can insert the necessary
compact information rows to do frame pointer based unwinding. The code
will be larger than 2**16 bytes in most real life programs, where the
previous implementation of `to_pages` would produce incomplete data.

This new revised implementation adds pages if the last page is "far
enough".

Test Plan
=========

Updated the tests and verified that unwinding large Go programs works
without any issues (see top of the stack / PR)