Fully support Zig
While profiling a Zig workload
(https://matklad.github.io/2025/05/06/performance-profile-visualization-challenge.html)
unwinding catastrophically fails, this is due to a couple of issues
documented on https://github.com/javierhonduco/lightswitch/issues/264.
This commit focuses on adding support for Zig executables compiled with
the Zig compiler before commit 130f7c2 as the first function,
`start` did not have the right DWARF unwind information.

In order to remediate this on our side, we attempt to recognise Zig
binaries with simple symbol-based heuristics, as there is no metadata of
any kind in the ELF files that we could use to know whether it was
compiled with the Zig compiler and if the version used contains this
bug.

The address range for the first function is read and the right unwind
information is then overriden for this file offset. We can't tell if the
compiler used has the bug or not until we aren't processing the unwind
information and given that all this is cached, the additional work of
always having traversing the symbol table is not a problem. I've verified that
this is the case with some self-profiles.

Notes
=====

- This is not enough to be able to successfully profile Tigerbeetle,
  more PRs coming, see issue and "Test Plan" there;
- The persisted unwind information for Zig should be reset. Will update
  the version in another PR as
https://github.com/javierhonduco/lightswitch/issues/266 needs a version
bump too